pyproject.toml
[tool.poetry]
name = "aqtra-docs"
version = "0.1.0"
description = "Documentation site (MkDocs Material)"
authors = ["Aqtra Team <team@example.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
mkdocs = ">=1.5,<1.7"
mkdocs-material = ">=9.5,<10"
mkdocs-glightbox = ">=0.4,<1"
mkdocs-mermaid2-plugin = ">=0.6,<1"
mkdocs-redirects = ">=1.2,<2"
mkdocs-git-revision-date-localized-plugin = ">=1.2,<2"
mkdocstrings = ">=0.30,<1"
mkdocs-static-i18n = ">=1.3,<2"
polib = ">=1.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7"
yamllint = "^1.35"

[tool.poetry.scripts]
docs-serve = "mkdocs.__main__:cli"
docs-build = "mkdocs.__main__:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
